Extract facilitators from slide to README on Japan
This commit adds facilitators list to README.md on Japan event.
As of now, facilitators list is included in workshop slide file.

It's good to extract informations which depends on each event to
README.md, because that workshop slide for Japan event was moved
to `japan/assets` folder as common asset.